+++ lib/src/instrument.rs
@@ -333,8 +333,10 @@ mod tests {
     const SAMPLE_RATE: u32 = 44_100;
 
     lazy_static! {
-        static ref BANK_A: [Wavetable<'static>; 1] =
-            [Wavetable::new(&waveform::saw::SAW_FACTORS, SAMPLE_RATE)];
+        static ref BANK_A: [Wavetable<'static>; 1] = [Wavetable::new(
+            &waveform::akwf::PERFECT_3_FACTORS,
+            SAMPLE_RATE
+        )];
         static ref WAVETABLE_BANKS: [&'static [Wavetable<'static>]; 1] = [&BANK_A[..]];
     }