+++ lib/src/waveform/mod.rs
@@ -1,6 +1,2 @@
 pub mod akwf;
-pub mod pulse;
-pub mod saw;
-pub mod sinc;
 pub mod sine;
-pub mod triangle;